Makes it ready to draw something on screen!
Adds Mesh structure which is responsible for initializing the proper
shader locations and structure that must be set so that the buffer is
able to send its data according to what it needs.

We aim:

--------                    --------
|shader|                    |Buffer|
--------                    --------
    |                           |
Material                    Geometry
    |                           |
    --------------Mesh-----------

Refactors other components to explicitly receive the opengl context

Refactors webpack.config so that we keep example in the example libs
folder.