Adds debug utils and lets more than one drawing
Debugging WebGL and OpenGL in general is hard. In this commit i add
an utility library that is widely used across the webgl dev env.

I also refactor the code a bit to support drawing multiple objects.
This is not optimal for same-type-objects-rendering as we loose a
bunch of perf with bindings. In the future we might optimize this
with SubArrays