inits a renderer-material-mesh-geom based render
This is the first commit over a major change that will happen in the
framework regarding how we put things on the screen. It tries to
encapsulate things in a logical manner in respect to how we prepare
the buffers and shaders without letting the user touch these low
level stuff.

Much more work needs to be done regarding materials, positioning,
camera and other stuff.