Updates Cube to new geometry interface and new ex
There was a problem regarding both view and projections. This commmit
finally adresses it by using the proper arguments to lookAt. It also
adds a comment about how the lookAt matrix is built.

Notice that the controls for the square example are diffente from those
used in this example. In the first we move the camera, while in this,
the cube per se.