Adds naive color handling for BasicMaterial
This commit adds a naive implementation of color handling through
materials and renderables. The way it's being done is pretty bad as it
relies on two separate buffers: 1 for vertices and another for colors,
having the same size, etc. We must wrap all of this in a single buffer
so that we reduce the amount of bindings, which is a pretty expensive
operation to be performed.

We can care about these when adding normals, which is the next step.