Gives shader the responsability for creating vbo
This commit makes a considerable improvement regarding speed as we cut
in 3 the number of bindings that we'd generaly perform for each object
in the scene (binding for vertex, color and normals ==> SHIT!). Now
it's much better. Hope this dont screws everything.

Also, another thing that we are going to be able to do is only send
those matrices that actually matter for a given shader. For example,
for those objects that do not rely on normals we might be able to just
pass a uniform that contains the PROJECTION*VIEW*MODEL matrix.

COOL!