Adds phong shading to normals material
This commits introduces phong shading local illumination to the normals
material. It also improves the way we render things by allowing each
shader to specify its normals programatically. After the positioning fix
we will be able to minimize the amount of matrix calculations that
happen in particular shaders because of this fix.