Starts the development of Body
The public API that i aim introduce here is a pretty simple one:
- position:
    Where in the global space this body is positioned
- scale:
    By which factor it is scaled
- at:
    Where is this body looking at?

Internally there's a _rotation quaternion that gets update, a
_modelMatrix and a _normalMatrix that are used by the framework. Also,
updates to those public things are done by direct setting of values.

Updates to the _normalMatrix and _modelMatrix are done in batch (every
set paints the internal state to dirty).