Updates camera positioning interface and demos
Even though i started thinking that the camera should inherit from Body
class just like the renderable, i didn't do it. The reason way i
stepped back is: the camera doesn't hold a model and a normal matrix
like the other bodies. Even though we can think of the viewMatrix as
some kind of translation+rotation etc think, it's a bit different and
it makes no sense to hold such unused structures for the camera.

Regarding the demos, the strange effect that it was being generated was
due to a bad setting of the 'at' vector (the orientation of the camera).
We were letting it fixed to the center of the scene while moving around.
We actually needed to update that vector so that when passing by the
position indicated by it the camera would not rotate.

note: the behavior IS what we'd expect as the direction during the
      lookAt calculation is obtained by performing a subtraction:
        * zAxis = target - camera.position (or something like this).