Separates cameras, initializes points and dyn buff
If we wish to do some 2D stuff we need to get an orthographic camera as
it doesn't make sense to rely on a perspective setting for 2D. As this
type of camera is pretty cool for performing some kinds of editting when
you don't want the distortion in size that perspective provokes.

We'll need to create some kind of 'dynamic buffer' that will deal with
dynamic changes to the buffer (we'll allow the user to change the position
of the control points).

Here i also start the Curve utility that will firstly implement BSplines.