Adds first NURBS implementation; Separates curves
This commits adds the first NURBS curves implementation. NURBS stands
for Non-uniform ration basis spline (see
https://en.wikipedia.org/wiki/Non-uniform_rational_B-spline). It's a
improvement over BSpline curves as it provides us a better control over
the 'power' of a control point by letting the user provide a weight for
each control point.

NURBS also lets us draw any kind of conic section through the capability
of weighting individual points. Now we're able to define closed forms
that can even describe a circle (exactly).

This commit DOES NOT include an example of a perfect simple.

Aside from the NURBS implementation it also adds a base Curve class and
create the initial file for the RaGs curve, which is the next step.